nir/xfb: fixing the xfb_offset for arrays of structs
GLSLang computes the xfb_offset for struct members. In fact, for basic
structs, the xfb nir gathering pass expect those to be filled, as one
struct variable is splitted to several nir variables, and those need
to have the xfb offset already set. See [1].

But in addition to that, after computing the xfb offset, GLSLang
unsets the xfb offset for the full struct, or array of structs.

This affects the case of variables that are array of structs, as they
are exposed as just one nir variable output. This commit resets the
xfb_offset for the nir_variable if any of the members has a xfb_offset
assigned.

FWIW, this was not needed for array of blocks because on that case we
are iteraring through all the members of the array of blocks, in order
to support the case where just some members were assigned a
xfb_offset.

[1] https://github.com/KhronosGroup/glslang/pull/154

Equivalent to previous patch (so comments applies), but implemented on
a different place. We would need to chose in which one.