spirv/nir: only set nir interface_type for block input/outputs
While processing input/output variables, local interface_type (used
for thing like knowing when split the variable) is used as default
value for nir variable interface type. But we shouldn't do this in
general, as on nir, interface_type is also used to know if we are on
an interface block. From nir.h:

   /**
    * For variables that are in an interface block or are an instance of an
    * interface block, this is the \c GLSL_TYPE_INTERFACE type for that block.
    *
    * \sa ir_variable::location
    */

This commit sets the interface_type only if it is an block.

Additionally, this commit sets the nir interface block if we have an
array of blocks. Previously such interface type were only took into
account for the cases were we were interested on splitting the
interface type. From the nir pov, we want to set the interface type
for arrays of blocks always.

v2: simplify how to assign the nir interface_type, using
    vtn_type_contains_block (Jason)